src/requirements.txt
python-dotenv==1.1.1
Flask==3.1.2
